Added register

+++ rest/tests.py
@@ -3,10 +3,11 @@ import os
 
 
 def test_dao_init():
+    if os.path.exists(ProfileDao.DB_PATH):
+        os.remove(ProfileDao.DB_PATH)
     ProfileDao()
     assert os.path.exists(ProfileDao.DB_PATH)
-    assert len(open(ProfileDao.DB_PATH).read()) > 0
-    os.remove(ProfileDao.DB_PATH)
+    assert len(open(ProfileDao.DB_PATH, 'rb').read()) > 0
 
 
 BASE_PROFILE = Profile(
@@ -23,18 +24,33 @@ BASE_PROFILE = Profile(
 )
 
 
+def test_dao_get_all_logins():
+    if os.path.exists(ProfileDao.DB_PATH):
+        os.remove(ProfileDao.DB_PATH)
+    dao = ProfileDao()
+    dao.insert_profile(BASE_PROFILE)
+
+    assert dao.get_all_logins() == [BASE_PROFILE.login]
+    profile = Profile(**vars(BASE_PROFILE))
+    profile.login = 'someone'
+    dao.insert_profile(profile)
+
+    assert set(dao.get_all_logins()) == {BASE_PROFILE.login, profile.login}
+
+
 def test_dao_get_profile():
+    if os.path.exists(ProfileDao.DB_PATH):
+        os.remove(ProfileDao.DB_PATH)
     dao = ProfileDao()
     dao.insert_profile(BASE_PROFILE)
 
     dao_profile = dao.lookup_profile(BASE_PROFILE.login)
     assert vars(dao_profile) == vars(BASE_PROFILE)
 
-    del dao
-    os.remove(ProfileDao.DB_PATH)
-
 
 def test_dao_modify_profile():
+    if os.path.exists(ProfileDao.DB_PATH):
+        os.remove(ProfileDao.DB_PATH)
     dao = ProfileDao()
     dao.insert_profile(BASE_PROFILE)
 
@@ -45,11 +61,10 @@ def test_dao_modify_profile():
     new_profile.mail, new_profile.gender = BASE_PROFILE.mail, BASE_PROFILE.gender
     assert vars(new_profile) == vars(BASE_PROFILE)
 
-    del dao
-    os.remove(ProfileDao.DB_PATH)
-
 
 def test_dao_finish_game():
+    if os.path.exists(ProfileDao.DB_PATH):
+        os.remove(ProfileDao.DB_PATH)
     dao = ProfileDao()
     dao.insert_profile(BASE_PROFILE)
 
@@ -66,6 +81,3 @@ def test_dao_finish_game():
     assert profile.session_count == BASE_PROFILE.session_count + 2
     assert profile.win_count == BASE_PROFILE.win_count + 1
     assert profile.lose_count == BASE_PROFILE.lose_count + 1
-
-    del dao
-    os.remove(ProfileDao.DB_PATH)
